Codebase cleanup & added imports
