+++ package.json
@@ -19,10 +19,11 @@
   "dependencies": {
     "awesome_starter": "1.0.6",
     "axios": "0.17.1",
+    "esm": "^3.0.77",
     "node-cron": "1.2.1",
-    "qrcode-terminal": "0.11.0",
     "node-telegram-bot-api": "0.30.0",
     "pm2": "3.0.3",
+    "qrcode-terminal": "0.11.0",
     "speakeasy": "2.0.0",
     "tail": "1.2.3"
   },
