+++ src/lib/blockheights.js
@@ -1,9 +1,9 @@
-const awesome = require("awesome_starter");
-const axios = require("axios");
-const settings = require("../config");
-const { bot } = require("./telegram");
+import awesome from "awesome_starter";
+import axios from "axios";
+import settings from "../config";
+import { bot } from "./telegram";
 
-const getBlockHeight = async node => {
+export const getBlockHeight = async node => {
   const blockHeightData = await axios
     .get(node)
     .catch(() =>
@@ -20,7 +20,7 @@ const getBlockHeight = async node => {
 
 //used for the cronjob
 //TODO: Fix potential duplicated code with respondBlockHeights()
-const compareBlockHeightsCron = async () => {
+export const compareBlockHeightsCron = async () => {
   const localBlockheight = await getBlockHeight(
     settings.localNodeURL
   ).catch(() =>
@@ -31,8 +31,7 @@ const compareBlockHeightsCron = async () => {
   let matchingBlockHeights = 0;
 
   if (localBlockheight === "0")
-    return bot.sendMessage(
-      settings.chatId,
+    return bot.reply(
       "I didn't manage to get the local block height during the routine check!"
     );
 
@@ -59,26 +58,23 @@ const compareBlockHeightsCron = async () => {
     console.fail(
       "Detected issue: The block heights are not matching with the rest of the network"
     );
-    await bot.sendMessage(
-      settings.chatId,
+    await bot.reply(
       `⚠️ Detected issue while executing the routine check`,
       { parse_mode: "HTML" }
     );
-    await bot.sendMessage(
-      settings.chatId,
+    await bot.reply(
       `<b>Local block height</b> is: ${localBlockheight}`,
       { parse_mode: "HTML" }
     );
-    await bot.sendMessage(settings.chatId, otherBlockMessage, { parse_mode: "HTML" });
-    await bot.sendMessage(
-      settings.chatId,
+    await bot.reply(otherBlockMessage, { parse_mode: "HTML" });
+    await bot.reply(
       "⚠️ Detected issue: The block heights are not matching with the rest of the network"
     );
   }
 };
 
 // Used for replying to user message
-const respondBlockHeights = async () => {
+export const respondBlockHeights = async () => {
   const localBlockheight = await getBlockHeight(
     settings.localNodeURL
   ).catch(() =>
@@ -89,19 +85,17 @@ const respondBlockHeights = async () => {
   let matchingBlockHeights = 0;
 
   if (localBlockheight === "0")
-    return bot.sendMessage(
-      settings.chatId,
+    return bot.reply(
       "Error getting the local block height"
     );
 
   console.log("Local block height is: ", localBlockheight);
-  bot.sendMessage(
-    settings.chatId,
+  bot.reply(
     `<b>Local block height</b> is: ${localBlockheight}`,
     { parse_mode: "HTML" }
   );
 
-  await bot.sendMessage(settings.chatId, `Parsing remote nodes ✨🔎`);
+  await bot.reply(`Parsing remote nodes ✨🔎`);
 
   for (let node of settings.remoteNodes) {
     let remoteNode = Object.assign({}, node);
@@ -119,7 +113,7 @@ const respondBlockHeights = async () => {
   }
 
   //publishing summary other block heights
-  await bot.sendMessage(settings.chatId, otherBlockMessage, {
+  await bot.reply(otherBlockMessage, {
     parse_mode: "HTML"
   });
 
@@ -130,21 +124,15 @@ const respondBlockHeights = async () => {
 
   if (sameBlockHeights) {
     console.success("Your node seems to have the correct height");
-    bot.sendMessage(
-      settings.chatId,
+    bot.reply(
       "👌 Your node seems to have the correct height"
     );
   } else {
     console.fail(
       "The block heights are not matching with the rest of the network"
     );
-    bot.sendMessage(
-      settings.chatId,
+    bot.reply(
       "😡 The block heights are not matching with the rest of the network"
     );
   }
 };
-
-exports.getBlockHeight = getBlockHeight;
-exports.compareBlockHeightsCron = compareBlockHeightsCron;
-exports.respondBlockHeights = respondBlockHeights;
